extended rate display
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@19058 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/shared/wlutils.h
@@ -103,7 +103,7 @@ extern int has_ht40(char *prefix);
 struct site_survey_list {
 	char SSID[33];
 	unsigned char BSSID[18];
-	uint8 channel;		/* Channel no. */
+	uint16 channel;		/* Channel no. */
 	uint16 frequency;		/* Frequency i.e. for superchannel */
 	int16 RSSI;		/* receive signal strength (in dBm) */
 	int16 phy_noise;	/* noise (in dBm) */
